+++ worker/src/RTC/WebRtcTransport.cpp
@@ -38,7 +38,10 @@ namespace RTC
 	  const std::string& id,
 	  RTC::Transport::Listener* listener,
 	  const FBS::WebRtcTransport::WebRtcTransportOptions* options)
-	  : RTC::Transport::Transport(shared, id, listener, options->base())
+	  // SCTP over WebRtcTransport runs within a DTLS session, so State Cookie
+	  // authentication is not needed.
+	  : RTC::Transport::Transport(
+	      shared, id, listener, options->base(), /*requireSctpStateCookieAuthentication*/ false)
 	{
 		MS_TRACE();
 
@@ -284,7 +287,10 @@ namespace RTC
 	  WebRtcTransportListener* webRtcTransportListener,
 	  const std::vector<RTC::ICE::IceCandidate>& iceCandidates,
 	  const FBS::WebRtcTransport::WebRtcTransportOptions* options)
-	  : RTC::Transport::Transport(shared, id, listener, options->base()),
+	  // SCTP over WebRtcTransport runs within a DTLS session, so State Cookie
+	  // authentication is not needed.
+	  : RTC::Transport::Transport(
+	      shared, id, listener, options->base(), /*requireSctpStateCookieAuthentication*/ false),
 	    webRtcTransportListener(webRtcTransportListener),
 	    iceCandidates(iceCandidates)
 	{